+++ src/components/blocks/article-slide/article-slide.tsx
@@ -1,7 +1,5 @@
 import { IconArrowRight } from "@/assets/icons";
-import Button from "@/components/button/button";
-import ArticleCard from "@/components/cards/article/article-card";
-import Slider from "@/components/slider/slider";
+import { ArticleCard, Button, Slider } from "@/components";
 import cn from "classnames";
 import { FC } from "react";
 import { SwiperSlide } from "swiper/react";
@@ -37,9 +35,11 @@ export const ArticleSlide: FC<ArticleSlideProps> = ({ className, children }) => 
           <ArticleCard productCode={5190} text={""} imageUrl={""} link={""} />
         </SwiperSlide>
       </Slider>
-      <Button className={cn(styles['button-mobile'],"color-white")} appearance="yellow">
-        Показать все <IconArrowRight />
-      </Button>
+      <div className={styles["button-mobile"]}>
+        <Button className={"color-white"} appearance="yellow">
+          Показать все <IconArrowRight />
+        </Button>
+      </div>
     </div>
   );
 };
