+++ src/components/blocks/contacts-map/contacts-map.props.ts
@@ -1,6 +1,6 @@
 import { HTMLAttributes } from "react";
 
-export interface MapCardProps extends HTMLAttributes<HTMLDivElement> {
+export interface ContactsMapProps extends HTMLAttributes<HTMLDivElement> {
   orient: "row" | "column";
   title: string;
   address: string;
