+++ src/components/blocks/home-intro/home-intro.props.ts
@@ -1,5 +1,5 @@
 import { HTMLAttributes } from "react";
 
-export interface IntroProps extends HTMLAttributes<HTMLDivElement> {
+export interface HomeIntroProps extends HTMLAttributes<HTMLDivElement> {
   data: [{ id: number; title: string; subtitle: string; caption?: string }];
 }
