+++ src/components/blocks/home-products/home-products.module.scss
@@ -24,6 +24,12 @@
   }
 }
 .item {
+  @include table {
+    display: none;
+    &:nth-child(-n + 6) {
+      display: grid;
+    }
+  }
 }
 
 .button {
