+++ src/components/blocks/home-products/home-products.props.ts
@@ -1,0 +1,3 @@
+import { HTMLAttributes } from "react";
+
+export interface HomeProductProps extends HTMLAttributes<HTMLDivElement> {}
