--- src/components/blocks/products/products-block.props.ts
@@ -1,3 +1,0 @@
-import { HTMLAttributes } from "react";
-
-export interface ProductBlockProps extends HTMLAttributes<HTMLDivElement> {}
